make ex_doc only-dev dep
Hello!
Your project looks neat :)

Now from my experience this little change is going to make life easier.

Generating a documentation is usually task you do on your dev machine. It's not necessary to carry this package as a prod subdep when someuse uses your package as a dep.

Also, a single git dep can screw deps resolution - again, this is usually not hurting when you're inside *this* project, but starts to be annoying when people grab your code from Hex as a dep.