Improve runtime performance via fewer calls
We'll pre-compute and pass around some values which we are highly likely
to need more than once. There's very little overhead in the use once
case. I suppose if it is common to call these and not use the
functions.values ever then this might be more expensive.

I have no idea what kind of side-effects might be expected, though

Some of this code was already done in this style, so I assume this is
readable enough and perhaps welcome.